+++ scripts/ci/libraries/_docker.env
@@ -24,9 +24,13 @@ HOST_OS
 HOST_HOME
 HOST_AIRFLOW_SOURCES
 PYTHON_MAJOR_MINOR_VERSION
+BACKEND
 VERSION_SUFFIX_FOR_PYPI
 VERSION_SUFFIX_FOR_SVN
 PRINT_INFO_FROM_SCRIPTS
 CI
 LOAD_DEFAULT_CONNECTIONS
 LOAD_EXAMPLES
+GITHUB_REGISTRY_PULL_IMAGE_TAG
+POSTGRES_VERSION
+MYSQL_VERSION
