+++ scripts/in_container/run_ci_tests.sh
@@ -31,11 +31,58 @@ set +x
 if [[ "${RES}" == "0" && ${CI:="false"} == "true" ]]; then
     echo "All tests successful"
     cp .coverage /files
+elif [[ "${RES}" != "0" ]]; then
+    EXTRA_ARGS=""
+    if [[ ${BACKEND} == "postgres" ]]; then
+        EXTRA_ARGS="--postgres-version ${POSTGRES_VERSION} "
+    elif [[ ${BACKEND} == "mysql" ]]; then
+        EXTRA_ARGS="--mysql-version ${MYSQL_VERSION} "
+    fi
+
+    >&2 echo "***********************************************************************************************"
+    >&2 echo "*"
+    >&2 echo "* ERROR! Some tests failed, unfortunately. Those might be transient errors,"
+    >&2 echo "*        but usually you have to fix something."
+    >&2 echo "*        See the above log for details."
+    >&2 echo "*"
+    >&2 echo "***********************************************************************************************"
+    >&2 echo "*  You can easily reproduce the failed tests on your dev machine/"
+    >&2 echo "*"
+    >&2 echo "*   When you have the source branch checked out locally:"
+    >&2 echo "*"
+    >&2 echo "*     Run all tests:"
+    >&2 echo "*"
+    >&2 echo "*       ./breeze --backend ${BACKEND} ${EXTRA_ARGS}--python ${PYTHON_MAJOR_MINOR_VERSION} --db-reset --test-type ${TEST_TYPE}  tests"
+    >&2 echo "*"
+    >&2 echo "*     Enter docker shell:"
+    >&2 echo "*"
+    >&2 echo "*       ./breeze --backend ${BACKEND} ${EXTRA_ARGS}--python ${PYTHON_MAJOR_MINOR_VERSION} --db-reset --test-type ${TEST_TYPE}  shell"
+    >&2 echo "*"
+    if [[ ${GITHUB_REGISTRY_PULL_IMAGE_TAG=} != "" ]]; then
+        >&2 echo "*   When you do not have sources:"
+        >&2 echo "*"
+        >&2 echo "*     Run all tests:"
+        >&2 echo "*"
+        >&2 echo "*      ./breeze --github-image-id ${GITHUB_REGISTRY_PULL_IMAGE_TAG} --backend ${BACKEND} ${EXTRA_ARGS}--python ${PYTHON_MAJOR_MINOR_VERSION} --db-reset --test-type ${TEST_TYPE} tests"
+        >&2 echo "*"
+        >&2 echo "*     Enter docker shell:"
+        >&2 echo "*"
+        >&2 echo "*      ./breeze --github-image-id ${GITHUB_REGISTRY_PULL_IMAGE_TAG} --backend ${BACKEND} ${EXTRA_ARGS}--python ${PYTHON_MAJOR_MINOR_VERSION} --db-reset --test-type ${TEST_TYPE} shell"
+        >&2 echo "*"
+    fi
+    >&2 echo "*"
+    >&2 echo "*   NOTE! Once you are in the docker shell, you can run failed test with:"
+    >&2 echo "*"
+    >&2 echo "*            pytest [TEST_NAME]"
+    >&2 echo "*"
+    >&2 echo "*   You can copy the test name from the output above"
+    >&2 echo "*"
+    >&2 echo "***********************************************************************************************"
 fi
 
 MAIN_GITHUB_REPOSITORY="apache/airflow"
 
-if [[ ${ONLY_RUN_QUARANTINED_TESTS:=} = "true" ]]; then
+if [[ ${TEST_TYPE:=} == "Quarantined" ]]; then
     if [[ ${GITHUB_REPOSITORY} == "${MAIN_GITHUB_REPOSITORY}" ]]; then
         if [[ ${RES} == "1" || ${RES} == "0" ]]; then
             echo
@@ -47,15 +94,7 @@ if [[ ${ONLY_RUN_QUARANTINED_TESTS:=} = "true" ]]; then
             echo "Pytest exited with ${RES} result. NOT Updating Quarantine Issue!"
             echo
         fi
-    else
-        echo
-        echo "GitHub repository '${GITHUB_REPOSITORY}'. NOT Updating Quarantine Issue!"
-        echo
     fi
-else
-    echo
-    echo "Regular tests. NOT Updating Quarantine Issue!"
-    echo
 fi
 
 if [[ ${CI:=} == "true" ]]; then
