+++ tests/core/test_config_templates.py
