+++ tests/core/test_configuration.py
