+++ tests/core/test_core.py
