+++ tests/core/test_local_settings.py
