+++ tests/core/test_sentry.py
