+++ tests/jobs/test_local_task_job.py
@@ -175,7 +175,7 @@ class TestLocalTaskJob(unittest.TestCase):
                 delta = (time2 - time1).total_seconds()
                 self.assertAlmostEqual(delta, job.heartrate, delta=0.05)
 
-    @pytest.mark.xfail(condition=True, reason="This test might be flaky in postgres/mysql")
+    @pytest.mark.quarantined
     def test_mark_success_no_kill(self):
         """
         Test that ensures that mark_success in the UI doesn't cause
@@ -360,6 +360,7 @@ class TestLocalTaskJob(unittest.TestCase):
         self.assertNotIn('reached_end_of_sleep', data,
                          'Task should not have been allowed to run to completion')
 
+    @pytest.mark.quarantined
     def test_mark_success_on_success_callback(self):
         """
         Test that ensures that where a task is marked suceess in the UI
