+++ tests/www/test_views.py
@@ -56,7 +56,6 @@ from airflow.ti_deps.dependencies_states import QUEUEABLE_STATES, RUNNABLE_STATE
 from airflow.utils import dates, timezone
 from airflow.utils.log.logging_mixin import ExternalLoggingMixin
 from airflow.utils.session import create_session
-from airflow.utils.sqlalchemy import using_mysql
 from airflow.utils.state import State
 from airflow.utils.timezone import datetime
 from airflow.utils.types import DagRunType
@@ -2388,7 +2387,7 @@ class TestTriggerDag(TestBase):
         self.assertIn('/trigger?dag_id=example_bash_operator', resp.data.decode('utf-8'))
         self.assertIn("return confirmDeleteDag(this, 'example_bash_operator')", resp.data.decode('utf-8'))
 
-    @pytest.mark.xfail(condition=using_mysql, reason="This test might be flaky on mysql")
+    @pytest.mark.quarantined
     def test_trigger_dag_button(self):
 
         test_dag_id = "example_bash_operator"
@@ -2404,7 +2403,7 @@ class TestTriggerDag(TestBase):
         self.assertIn(DagRunType.MANUAL.value, run.run_id)
         self.assertEqual(run.run_type, DagRunType.MANUAL.value)
 
-    @pytest.mark.xfail(condition=using_mysql, reason="This test might be flaky on mysql")
+    @pytest.mark.quarantined
     def test_trigger_dag_conf(self):
 
         test_dag_id = "example_bash_operator"
